split goal in trigger area and actual goal
