+++ Assets/Scripts/Goal.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dc81897e7e830cf46a8911c1fc9ea537
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
