+++ Assets/Scripts/Goal/GoalArea.cs
@@ -1,0 +1,30 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.Events;
+
+public class GoalArea : MonoBehaviour
+{
+    public static UnityEvent PlayerInArea = new UnityEvent();
+
+    private void OnTriggerEnter(Collider other)
+    {
+        print("Player is near Goal");
+        if (other.GetComponent<CharacterController>())
+        {
+            PlayerInArea?.Invoke();
+
+            // turn on basic effects here
+        }
+    }
+
+    private void OnTriggerExit(Collider other)
+    {
+        if (other.GetComponent<CharacterController>())
+        {
+            PlayerInArea?.Invoke();
+
+            // turn off basic effects here
+        }
+    }
+}
