+++ Assets/Scripts/Goal/GoalArea.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e9e6ed0cef48d36428838ea789dda0c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
