+++ Assets/Scripts/Goal/GoalObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e46ef98cf75a78e41ba58ecf43a729b2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
